Temp jitpack for testing purposes
I do not intend to distribute a modularized version of jnigen, nor do I intend to infringe on the credit which should rightfully be with its original creators. What I do intend to do, is to fuck around and find out.
This jitpack setup is PURELY for Melange to test if it can be modularized if jnigen is.